+++ WHATSNEW.md
@@ -1,7 +1,9 @@
 # Version 2022.2
+* Updated to [MAME 243](https://www.mamedev.org/releases/whatsnew_0243.txt).
 * hi-res font for `MAME` Config Menu.
 * use the entire screen (ignore safe area) if `Keep Aspect Ratio` is `OFF`
-
+* better touch mouse support, tap is LEFT button, two finger tap is RIGHT button.
+    - touch mouse can be enabled or disabled in `Settings` > `Input Options` > `Touch Analog`
 
 # Version 2022.1
 * Updated to [MAME 241](https://www.mamedev.org/releases/whatsnew_0241.txt).