+++ WHATSNEW.md
@@ -1,5 +1,5 @@
 # Version 2022.2
-* Updated to [MAME 243](https://www.mamedev.org/releases/whatsnew_0243.txt).
+* Updated to [MAME 244](https://www.mamedev.org/releases/whatsnew_0244.txt).
 * hi-res font for `MAME` Config Menu.
 * use the entire screen (ignore safe area) if `Keep Aspect Ratio` is `OFF`
 * better touch mouse support, tap is LEFT button, two finger tap is RIGHT button.